# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Garment POS API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Authenticated }
  /inventory/products:
    get:
      summary: List products with variants and stock
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Product list }
    post:
      summary: Create product
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
  /inventory/variants/scan/{barcode}:
    get:
      summary: Scan barcode to get variant
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: barcode
          required: true
          schema: { type: string }
      responses:
        '200': { description: Variant }
        '404': { description: Not found }
  /billing/invoices:
    post:
      summary: Create invoice and decrement stock transactionally
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Invoice created }
  /customers:
    get:
      summary: List customers
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Customer list }
    post:
      summary: Create customer
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
  /purchase/suppliers:
    get:
      summary: List suppliers
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Supplier list }
    post:
      summary: Create supplier
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
  /cash/sessions:
    get:
      summary: List cash sessions
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Cash sessions }
  /cash/sessions/open:
    post:
      summary: Open new cash register session
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Opened }
  /reports/dashboard:
    get:
      summary: Dashboard KPIs
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: KPI payload }
  /print/receipt:
    post:
      summary: Print ESC/POS receipt
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Print queued }

  /sync/events:
    post:
      summary: Queue offline sync event
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Queued }
  /sync/events/{deviceId}:
    get:
      summary: Pull pending sync events
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Event list }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT